amazon clone project created
The Amazon Clone Project aims to replicate the look and feel of the Amazon homepage. This project is an excellent opportunity to practice and enhance my front-end development skills, particularly in HTML and CSS. The landing page has been designed with a keen eye for detail, using attractive images to elevate the user experience.